Initial websocket support
This adds initial websocket support.  It is not complete yet as while
you can create a websocket and send data via the websocket, receiving
data from the websocket does not work, for reasons I've yet to
determine.

This changes the API for route_request, because Reel::Request#connection
is not defined. It includes the request object in a 'reel.request' entry
in the env, so that rack apps can do:

  ws = env['reel.request'].websocket if env['reel.request'].websocket?